Add files via upload
This C Project is all about the good old game, "The 15 Puzzle" (Sliding Game). The Game is all about Solving the Puzzle by sliding the number slides and solving the puzzle in the least number of moves/slides. The project is fully developed in C, and all types of enhancements are Welcome. Please add My project to this repo.